Refactor client src structure
This commit resolve #50 issue.
