+++ client/src/components/game/GameDetail.vue
