+++ client/src/components/game/GameGrid.vue
