+++ client/src/components/game/GameInfo.vue
