+++ client/src/components/game/GameInfoExtra.vue
