+++ client/src/components/home/Home.vue
@@ -42,7 +42,7 @@
 
 <script>
 import { mapState, mapActions } from 'vuex';
-import AlertField from './AlertField.vue';
+import AlertField from '../util/AlertField.vue';
 import HomeAdvertise from './HomeAdvertise.vue';
 import HomeRanking from './HomeRanking.vue';
 import HomeNavi from './HomeNavi.vue';
