+++ client/src/components/home/HomeNavi.vue
