+++ client/src/components/info/Account.vue
@@ -50,7 +50,7 @@
 
 <script>
 import { mapState } from 'vuex';
-import AlertField from './AlertField.vue';
+import AlertField from '../util/AlertField.vue';
 
 export default {
   computed: {
