+++ client/src/components/rank/RankingAdvertise.vue
