+++ client/src/components/util/Infobar.vue
